refactor(init,upgrade): pivot squad_state MCP write to repo-root .mcp.json
Reverts iter-7's HOME-write to ~/.copilot/mcp-config.json. Copilot CLI
5.3+ auto-loads .mcp.json walking cwd up to git root, so writing the
squad_state entry into the repo root (or .squad subdir) is sufficient
and avoids polluting the user's global MCP config.

- Replace mcp-home.ts with mcp-root.ts (ensureSquadStateMcpInRoot +
  tombstoneStaleSquadStateInProjectMcp).
- Wire init.ts and upgrade.ts to the new module. Idempotent four-field
  match keeps user MCP entries untouched.
- Tombstone helper removes only the stale squad_state key from any
  pre-existing project mcp-config.json; never clobbers other servers.

Per Tamir's directive: `no writes to HOME, only repo or .squad subdir`.
